Introduce KeyExtractor strategy interface which contains both function and metadata to extract keys (IDs) from an entity
Remove `IdExtractor` and `IdResolver` interfaces to be replaced by single
composite `KeyExtractor` API.

Expose key AST (as Expression) to optimize certain queries. Currently useful
for Geode when query (OQL) can be replaced by key lookup (eg. `Region.getAll(...)`)

Provide several factory methods to lookup keys using reflection or use generic
java function to extract a key.

Change `*Setup` (eg. `GeodeSetup`) config classes to accept `KeyExtractor.Factory` instance.

+++ criteria/geode/src/org/immutables/criteria/geode/SyncDelete.java
@@ -49,10 +49,13 @@ class SyncDelete implements Callable<WriteResult> {
 
     Expression filter = operation.query().filter().orElseThrow(() -> new IllegalStateException("For " + operation));
 
-    IdOnlyFilter idOnly = new IdOnlyFilter(filter, session.idProperty);
-    // special case when expression contains only ID / key attribute
-    if (idOnly.hasOnlyIds()) {
-      return deleteByKeys(idOnly.toList());
+    // special case when expression may contain only ID / key attribute
+    // assume idProperty is defined (see IdResolver)
+    if (session.idProperty != null) {
+      IdOnlyFilter idOnly = new IdOnlyFilter(filter, session.idProperty);
+      if (idOnly.hasOnlyIds()) {
+        return deleteByKeys(idOnly.toList());
+      }
     }
 
     GeodeQueryVisitor visitor = new GeodeQueryVisitor(true, path -> String.format("e.value.%s", session.pathNaming.name(path)));
